[package][react-ui][refactor] restructure component stories
